Move optional feature type parsing to parser.ts
